chore(neon-build): Re-organize to prepare for changes in N-API build logic
* Update to 2018 edition
* Re-structure to platform module structure
* Remove `cfg-if`
* Make `ureq` unconditionally required
* Document `neon-build`
