+++ crates/neon-build/Cargo.toml
@@ -5,14 +5,9 @@ authors = ["Dave Herman <david.herman@gmail.com>"]
 description = "Build logic required for Neon projects."
 repository = "https://github.com/neon-bindings/neon"
 license = "MIT/Apache-2.0"
+edition = "2018"
 build = "build.rs"
 
 [dependencies]
 neon-sys = { version = "=0.5.3", path = "../neon-sys", optional = true }
-cfg-if = "0.1.9"
-
-[build-dependencies]
-cfg-if = "0.1.9"
-
-[target.'cfg(windows)'.dependencies]
 ureq = { version = "1.3.0", default-features = false, features = ["native-tls"] }
