+++ crates/neon-build/src/legacy/linux.rs
@@ -1,0 +1,1 @@
+pub(crate) fn setup() {}
