+++ crates/neon-build/src/napi/linux.rs
@@ -1,0 +1,1 @@
+pub(crate) fn setup() {}
