+++ crates/neon-build/src/napi/macos.rs
@@ -1,0 +1,4 @@
+pub(crate) fn setup() {
+    println!("cargo:rustc-cdylib-link-arg=-undefined");
+    println!("cargo:rustc-cdylib-link-arg=dynamic_lookup");
+}
